requirements.txt
cluster==1.4.1.post1
cycler==0.10.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
django-import-export==1.0.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
odfpy==1.3.6
openpyxl==2.5.3
psycopg2==2.7.6
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.2
six==1.11.0
sklearn==0.0
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==4.1
xlrd==1.1.0
xlwt==1.3.0
